WIP: TransactionRepositoryImpl test
